move styles into their own folder
