+++ src/assets/styles/footer.css
